Внес все правки
