+++ src/App.tsx
@@ -1,4 +1,4 @@
-import { GameAreaPage } from './page/GameAreaPage/GameAreaPage'
+import { GameAreaPage } from './page/game-area-page'
 
 export const App = () => {
   return <GameAreaPage />
