--- src/components/UI/AreaCard.tsx
@@ -1,7 +1,0 @@
-interface IAreaCardProps {
-
-}
-
-export const AreaCard = ({}: IAreaCardProps) => {
-  return <div className="areaCard">7</div>;
-};
