+++ src/components/UI/area-button/area-button.tsx
