+++ src/components/UI/area-card/araea-card.module.scss
@@ -1,0 +1,11 @@
+.areaCard {
+    border-radius: 50%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: #304859;
+    font-size: calc(1.9vw + 17px);
+    color: #fcfcfc;
+    font-weight: 700;
+    cursor: pointer;
+  }
