+++ src/components/UI/area-card/area-card.tsx
@@ -1,0 +1,6 @@
+import style from './araea-card.module.scss'
+interface IAreaCardProps {}
+
+export const AreaCard = ({}: IAreaCardProps) => {
+  return <div className={style.areaCard}>7</div>
+}
