+++ src/components/UI/area-card/index.ts
@@ -1,0 +1,1 @@
+export { AreaCard } from './area-card'
