+++ src/components/UI/play-card/index.ts
@@ -1,0 +1,1 @@
+export { PlayerCard } from './player-card'
