+++ src/components/UI/play-card/play-card.module.scss
@@ -1,0 +1,10 @@
+.playerCard {
+  padding: 24px 20px;
+  flex-grow: 1;
+  display: flex;
+  justify-content: space-between;
+  background-color: #dfe7ec;
+  border-radius: 10px;
+}
+
+
