+++ src/components/game-area-header/index.ts
@@ -1,0 +1,1 @@
+export { GameHeader } from './game-header'
