+++ src/components/game-area-main/GameMain.module.scss
@@ -1,0 +1,14 @@
+.gameAreaMain {
+  display: grid;
+  grid-template-columns: repeat(6, 80px);
+  grid-template-rows: repeat(6, 80px);
+  gap: calc(0.8vw + 5.2px);
+  width: fit-content;
+  margin: 0 auto;
+}
+@media (max-width: 750px) {
+  .gameAreaMain {
+    grid-template-columns: repeat(6, 48px);
+    grid-template-rows: repeat(6, 48px);
+  }
+}
