+++ src/components/game-area-main/game-area-main.tsx
@@ -1,10 +1,11 @@
-import { AreaCard } from './UI/AreaCard'
+import style from './GameMain.module.scss'
+import { AreaCard } from '../UI/area-card/index'
 
 interface IGameAreaMainProps {}
 
 export const GameAreaMain = ({}: IGameAreaMainProps) => {
   return (
-    <div className="gameAreaMain">
+    <div className={style.gameAreaMain}>
       <AreaCard />
       <AreaCard />
       <AreaCard />
