+++ src/components/game-footer/game-footer.module.scss
@@ -1,0 +1,6 @@
+footer {
+  display: flex;
+  justify-content: space-between;
+  gap: 30px;
+  
+}
