--- src/page/GameAreaPage/GameArea.scss
@@ -1,63 +1,0 @@
-.gameArea {
-  padding: calc(4vw + 8.8px) calc(13.2vw - 25.6px) calc(4.7vw + 6.4px);
-  background-color: #fcfcfc;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    .headerBtns {
-      display: flex;
-      gap: 16px;
-      button {
-        padding: 14px 28px;
-        border-radius: 28px;
-        border: 1px solid transparent;
-        font-size: 20px;
-        font-weight: 700;
-        color: #fcfcfc;
-        background-color: #fda214;
-        cursor: pointer;
-      }
-      button:last-child {
-        color: #304859;
-        background-color: #dfe7ec;
-      }
-    }
-  }
-  .gameAreaMain {
-    display: grid;
-    grid-template-columns: repeat(6, 80px);
-    grid-template-rows: repeat(6, 80px);
-    gap: calc(0.8vw + 5.2px);
-    width: fit-content;
-    margin: 0 auto;
-    .areaCard {
-      border-radius: 50%;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      background-color: #304859;
-      font-size: calc(1.9vw + 17px);
-      color: #fcfcfc;
-      font-weight: 700;
-      cursor: pointer;
-    }
-  }
-  footer {
-    display: flex;
-    justify-content: space-between;
-    gap: 30px;
-    .playerCard {
-      padding: 24px 20px;
-      flex-grow: 1;
-      display: flex;
-      justify-content: space-between;
-      background-color: #dfe7ec;
-      border-radius: 10px;
-    }
-  }
-}
