+++ src/page/game-area-page/GameArea.module.scss
@@ -1,0 +1,8 @@
+.gameArea {
+  padding: calc(4vw + 8.8px) calc(13.2vw - 25.6px) calc(4.7vw + 6.4px);
+  background-color: #fcfcfc;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
