+++ src/page/game-area-page/index.ts
@@ -1,0 +1,1 @@
+export { GameAreaPage } from './game-area-page'
